FIxed file locations
Fixed file location for User class
